holy shit fucking tablets!!!!!
(add NX and Android support (S1F only, you can build an Absolute switch binary but not Android yet. I have a WIP folder for this, but it won't take forever to set up and get ready. I will push it when we get full touch support in S2A))