Initial changes for macOS to build and run
No bundled dependencies required. Follow the steps in the instructions to download SDL2, libogg and libvorbis and extract them into the appropriate folders. You may also need to run the configure steps on ogg and vorbis for missing files to be generated, but once all files exist the LibOggAndVorbis Xcode project is a subproject of RSDKv4 so everything will build for you