Final mac autobuild setup
This is not ideal. Artifacts zips it too, making a double zip. But we need to zip it ourselves or the symlinks get lost and the app breaks. But the app is broken anyway because it's unsigned and you need to run "xattr -cr RSDKv4.app" for it to not say that it's damaged so it can run.